consolidate package structure
Currently, the structure is a complete mess:

    org.terasology
    ├── behaviors
    │   ├── actions
    │   ├── components
    │   └── system
    ├── flexiblemovement
    │   ├── actions
    │   ├── debug
    │   ├── node
    │   ├── plugin
    │   └── system
    └── minion
        └── move

I propose to move everything into the `org.terasology.module` namespace.

I propose to group everything under a commond `behaviors` package (the module's namespace).

Each of the sub-packages currently follows the separation by `actions`, `components`, and `systems`.
If we'd keep this structure, the target picture might look like this:

    org.terasology.module
    └──behaviors
        ├── behaviors
        ├── actions
        ├── components
        ├── debug
        ├── plugin
        └── system