fix(set_target_nearby_block): ensure target block is walkable
Without this check, we can end up with target blocks that are not walkable, screwing up the pathfinding logic. This small addition will fail the behavior action when the target is not walkable, in the hope that we'll just give it another try.

Several ideas on how to further improve this behavior action are added as TODO comments.