test: fix MovementTests by moving `testCharacter` prefab to top-level
There seems to be some issue with loading prefabs from nested
sub-folders in tests. For some reason, the AssetManager could not find
the `Behaviors:testCharacter` prefab which was located at
```
assets/prefabs/test/testCharacter.prefab
```
To fix this (for now) I moved the asset one level up, which seems
sufficient for the engine to find the prefab also for test cases.

In addition, I've added CoreAssets as dependency for the MTE tests, as
they are referencing blocks from that module when setting up the test
world.